+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Add new boosting decision routes.
 - Add boosting decision route to pandas client.
+- Add `craftai.pandas.client.get_generator_operations` to return a pandas DataFrame
 
 ## [2.1.1](https://github.com/craft-ai/craft-ai-client-python/compare/v2.1.0...v2.1.1) - 2021-02-25 ##